Refactor styles
Seperate each style into its own file for better organization
