+++ include/unwieldy-ui/apix.hpp
@@ -10,4 +10,4 @@
 
 #include "api.hpp"
 #include "fonts.hpp"
-#include "internal/styles.hpp"
+#include "styles.hpp"
