--- include/unwieldy-ui/internal/styles.hpp
@@ -1,26 +1,0 @@
-#pragma once
-#include "lvgl/lvgl.h"
-
-extern lv_style_t style_bg;
-extern lv_style_t style_transp;
-
-extern lv_style_t style_list;
-extern lv_style_t style_list_btn;
-extern lv_style_t style_list_btn_pr;
-
-extern lv_style_t style_btn;
-extern lv_style_t style_btn_primary;
-extern lv_style_t style_btn_primary_pr;
-extern lv_style_t style_btn_outline;
-extern lv_style_t style_btn_outline_pr;
-
-extern lv_style_t style_text_mono;
-extern lv_style_t style_text_small;
-extern lv_style_t style_text_medium;
-extern lv_style_t style_text_large;
-extern lv_style_t style_text_centered;
-
-extern lv_style_t style_bar_button;
-extern lv_style_t style_bar_button_pr;
-extern lv_style_t style_bar_list;
-extern lv_style_t style_bar_bg;
