+++ src/unwieldy-ui/styles/initialize.cpp
@@ -1,0 +1,10 @@
+#include "unwieldy-ui/styles.hpp"
+
+void gui::theme::_initialize() {
+	_init_colors();
+	_init_style_misc();
+	_init_style_text();
+	_init_style_btn();
+	_init_style_list();
+	_init_style_bar();
+}
